Implement GtkWidget::multidevice-event.
Callbacks for this signal will receive a GtkDeviceGroup and a
GtkMultiDeviceEvent, which groups together the latest motion events for the
devices contained in a group.

GtkMultiDeviceEvent also has a pointer to the last updated event (the one that
triggered the signal to be emitted) and a hint about whether there was a new
device added, removed (updated_event will be NULL) or updated.